Refactor knowledge pipeline into modules
