+++ pipeline/__init__.py
@@ -1,0 +1,4 @@
+from .config import PipelineConfig
+from .core import KnowledgePipeline, cli
+
+__all__ = ["PipelineConfig", "KnowledgePipeline", "cli"]
